- Add back InvocationUtil - Update comment.
