+++ core/src/main/java/com/linecorp/armeria/internal/metric/RequestMetricSupport.java
@@ -18,6 +18,7 @@ package com.linecorp.armeria.internal.metric;
 
 import static com.linecorp.armeria.common.metric.MoreMeters.newDistributionSummary;
 import static com.linecorp.armeria.common.metric.MoreMeters.newTimer;
+import static java.util.Objects.requireNonNull;
 
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
@@ -25,9 +26,12 @@ import java.util.concurrent.atomic.LongAdder;
 
 import javax.annotation.Nullable;
 
+import com.google.common.collect.ImmutableList;
+
 import com.linecorp.armeria.client.ClientConnectionTimings;
 import com.linecorp.armeria.client.ResponseTimeoutException;
 import com.linecorp.armeria.client.WriteTimeoutException;
+import com.linecorp.armeria.common.HttpStatus;
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.RpcResponse;
 import com.linecorp.armeria.common.logging.RequestLog;
@@ -39,6 +43,7 @@ import com.linecorp.armeria.server.RequestTimeoutException;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.DistributionSummary;
 import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.Timer;
 import io.netty.util.AttributeKey;
 
@@ -60,6 +65,22 @@ public final class RequestMetricSupport {
         ctx.log().addListener(log -> onRequest(log, meterIdPrefixFunction, server),
                               RequestLogAvailability.REQUEST_HEADERS,
                               RequestLogAvailability.REQUEST_CONTENT);
+    }
+
+    /**
+     * Appends {@link HttpStatus} to {@link Tag}.
+     */
+    public static void appendHttpStatusTag(ImmutableList.Builder<Tag> tagListBuilder, RequestLog log) {
+        requireNonNull(tagListBuilder, "tagListBuilder");
+        requireNonNull(log, "log");
+        // Add the 'httpStatus' tag.
+        final HttpStatus status;
+        if (log.isAvailable(RequestLogAvailability.RESPONSE_HEADERS)) {
+            status = log.status();
+        } else {
+            status = HttpStatus.UNKNOWN;
+        }
+        tagListBuilder.add(Tag.of("httpStatus", status.codeAsText()));
     }
 
     private static void onRequest(RequestLog log, MeterIdPrefixFunction meterIdPrefixFunction, boolean server) {
